Folder "package" path name must be lower case
